imgsearchapi/requirements.txt
###### Requirements with Version Specifiers ######`
fastapi==0.78.0 
uvicorn==0.17.6 
elasticsearch==8.2.0 
python-multipart==0.0.5
Pillow==9.1.1
torch==1.12.0
torchvision==0.13.0
pandas==1.4.2